requirements.txt
annotated-types==0.7.0
anyio==4.3.0
asgiref==3.8.1
boto3==1.34.109
botocore==1.34.109
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
cohere==5.5.0
Django==5.0.6
exceptiongroup==1.2.1
fastavro==1.9.4
filelock==3.14.0
fsspec==2024.5.0
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.0
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.0
idna==3.7
jmespath==1.0.1
packaging==24.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.1
rsa==4.9
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
tokenizers==0.19.1
tqdm==4.66.4
types-requests==2.32.0.20240521
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
